+++ .changeset/modern-roses-raise.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/coin-aptos": patch
+---
+
+Add listOperations to Aptos alpaca
