+++ libs/coin-modules/coin-aptos/src/__tests__/api/index.integ.test.ts
@@ -14,8 +14,8 @@ describe("createApi", () => {
   });
 
   const sender: AptosSender = {
-    xpub: "0x474dd8fad13de7ebc82e1cb7ec4e5320887a58010fc484ed5bc8c5ed73fcd8b0",
-    freshAddress: "0xa0d8abc262e3321f87d745bd5d687e8f3fb14c87d48f840b6b56867df0026ec8",
+    xpub: "0xd1a8c6a1cdd52dd40c7ea61ee4571fb51fcae440a594c1eca18636928f1d3956",
+    freshAddress: "0x445fa0013887abd1a0c14acdec6e48090e0ad3fed3e08202aac15ca14f3be26b",
   };
   const recipient: AptosSender = {
     xpub: "0x64159425ccc6e755b91dc801b93d182af978c4624d9064facaa9b147544db87f",
@@ -87,6 +87,53 @@ describe("createApi", () => {
 
       expect(balances.length).toBeGreaterThan(0);
       expect(balances[0].value).toBeGreaterThan(0);
+    });
+  });
+
+  describe("listOperations", () => {
+    it("returns operations from account", async () => {
+      const block = await api.lastBlock();
+
+      const [operations] = await api.listOperations(sender.freshAddress, {
+        minHeight: block.height,
+      });
+
+      expect(operations).toBeInstanceOf(Array);
+      expect(operations.length).toBeGreaterThanOrEqual(1);
+
+      const txINHash = "0xf8c8a486c8e0c0c530f92ea5b26220829e8f8e24f8b0d9f35b57dbd804d36daf";
+      const txOUTHash = "0xf980601fe40ad1dab0cc68fe08d2bc95c73e2a21c6d257475e0879394638058e";
+
+      const operationIN = operations.find(operation => operation.id === txINHash);
+      const operationOUT = operations.find(operation => operation.id === txOUTHash);
+
+      expect(operationIN).toMatchObject({
+        type: "IN",
+        value: 20000000n,
+        recipients: [sender.freshAddress],
+        senders: ["0xa0d8abc262e3321f87d745bd5d687e8f3fb14c87d48f840b6b56867df0026ec8"],
+        asset: { type: "native" },
+        tx: {
+          hash: "0xf8c8a486c8e0c0c530f92ea5b26220829e8f8e24f8b0d9f35b57dbd804d36daf",
+          block: { height: 0 },
+          fees: 1100n,
+          date: new Date("2025-03-11T16:27:53.180Z"),
+        },
+      });
+
+      expect(operationOUT).toMatchObject({
+        type: "OUT",
+        value: 119900n,
+        recipients: ["0xd20fa44192f94ba086ab16bfdf57e43ff118ada69b4c66fa9b9a9223cbc068c1"],
+        senders: [sender.freshAddress],
+        asset: { type: "native" },
+        tx: {
+          hash: "0xf980601fe40ad1dab0cc68fe08d2bc95c73e2a21c6d257475e0879394638058e",
+          block: { height: 0 },
+          fees: 99900n,
+          date: new Date("2024-12-18T14:14:59.703Z"),
+        },
+      });
     });
   });
 });
