+++ libs/coin-modules/coin-aptos/src/__tests__/network/client.test.ts
@@ -9,14 +9,16 @@ import {
   RawTransaction,
   Serializable,
   postAptosFullNode,
+  TransactionResponseType,
 } from "@aptos-labs/ts-sdk";
 import network from "@ledgerhq/live-network";
 import BigNumber from "bignumber.js";
 import { AptosAPI } from "../../network";
 import { AptosAsset, AptosExtra, AptosSender } from "../../types/assets";
-import { TransactionIntent } from "@ledgerhq/coin-framework/api/types";
+import { Pagination, TransactionIntent } from "@ledgerhq/coin-framework/api/types";
 import { TokenCurrency } from "@ledgerhq/types-cryptoassets";
 import { APTOS_ASSET_ID } from "../../constants";
+import { AptosTransaction } from "../../types";
 
 jest.mock("@aptos-labs/ts-sdk");
 jest.mock("@apollo/client");
@@ -27,6 +29,10 @@ let mockedHex: jest.Mocked<any>;
 
 jest.mock("@ledgerhq/live-network/network");
 const mockedNetwork = jest.mocked(network);
+
+jest.mock("@ledgerhq/cryptoassets/tokens", () => ({
+  findTokenByAddressInCurrency: jest.fn().mockReturnValue("token-name"),
+}));
 
 describe("Aptos API", () => {
   beforeEach(() => {
@@ -550,6 +556,7 @@ describe("Aptos API", () => {
       });
     });
   });
+
   describe("broadcast", () => {
     it("broadcasts the transaction", async () => {
       mockedPost.mockImplementation(async () => ({ data: { hash: "ok" } }));
@@ -649,6 +656,452 @@ describe("Aptos API", () => {
       expect(balances).toHaveLength(1);
       expect(balances[0].asset_type).toBe(assets[0].asset_type);
       expect(balances[0].amount).toStrictEqual(BigNumber(assets[0].amount));
+    });
+  });
+
+  describe("listOperations", () => {
+    it("list of operations", async () => {
+      const api = new AptosAPI("aptos");
+      const address = "0x12345";
+      const pagination: Pagination = { minHeight: 0 };
+
+      const txs: AptosTransaction[] = [
+        {
+          version: "2532591427",
+          hash: "0x3f35",
+          state_change_hash: "0xb480",
+          event_root_hash: "0x3fa1",
+          state_checkpoint_hash: null,
+          gas_used: "12",
+          success: true,
+          vm_status: "Executed successfully",
+          accumulator_root_hash: "0x319f",
+          changes: [
+            {
+              address: "0x4e5e",
+              state_key_hash: "0x3c0c",
+              data: {
+                type: "0x1::coin::CoinStore<0xd111::staked_coin::StakedAptos>",
+                data: {
+                  coin: {
+                    value: "4000000",
+                  },
+                  deposit_events: {
+                    counter: "9",
+                    guid: {
+                      id: {
+                        addr: "0x4e5e",
+                        creation_num: "4",
+                      },
+                    },
+                  },
+                  frozen: false,
+                  withdraw_events: {
+                    counter: "6",
+                    guid: {
+                      id: {
+                        addr: "0x4e5e",
+                        creation_num: "5",
+                      },
+                    },
+                  },
+                },
+              },
+              type: "write_resource",
+            },
+            {
+              address: address,
+              state_key_hash: "0x1709",
+              data: {
+                type: "0x1::coin::CoinStore<0x1::aptos_coin::AptosCoin>",
+                data: {
+                  coin: {
+                    value: "68254118",
+                  },
+                  deposit_events: {
+                    counter: "46",
+                    guid: {
+                      id: {
+                        addr: address,
+                        creation_num: "2",
+                      },
+                    },
+                  },
+                  frozen: false,
+                  withdraw_events: {
+                    counter: "89",
+                    guid: {
+                      id: {
+                        addr: address,
+                        creation_num: "3",
+                      },
+                    },
+                  },
+                },
+              },
+              type: "write_resource",
+            },
+            {
+              address: address,
+              state_key_hash: "0x5520",
+              data: {
+                type: "0x1::coin::CoinStore<0xd111::staked_coin::StakedAptos>",
+                data: {
+                  coin: {
+                    value: "1000000",
+                  },
+                  deposit_events: {
+                    counter: "7",
+                    guid: {
+                      id: {
+                        addr: address,
+                        creation_num: "10",
+                      },
+                    },
+                  },
+                  frozen: false,
+                  withdraw_events: {
+                    counter: "13",
+                    guid: {
+                      id: {
+                        addr: address,
+                        creation_num: "11",
+                      },
+                    },
+                  },
+                },
+              },
+              type: "write_resource",
+            },
+            {
+              address: address,
+              state_key_hash: "0x6f1e",
+              data: {
+                type: "0x1::account::Account",
+                data: {
+                  authentication_key: address,
+                  coin_register_events: {
+                    counter: "5",
+                    guid: {
+                      id: {
+                        addr: address,
+                        creation_num: "0",
+                      },
+                    },
+                  },
+                  guid_creation_num: "12",
+                  key_rotation_events: {
+                    counter: "0",
+                    guid: {
+                      id: {
+                        addr: address,
+                        creation_num: "1",
+                      },
+                    },
+                  },
+                  rotation_capability_offer: {
+                    for: {
+                      vec: [],
+                    },
+                  },
+                  sequence_number: "122",
+                  signer_capability_offer: {
+                    for: {
+                      vec: [],
+                    },
+                  },
+                },
+              },
+              type: "write_resource",
+            },
+            {
+              state_key_hash: "0x6e4b",
+              handle: "0x1b85",
+              key: "0x0619",
+              value: "0x1ddaf8da3b1497010000000000000000",
+              type: "write_table_item",
+            },
+          ],
+          sender: address,
+          sequence_number: "121",
+          max_gas_amount: "12",
+          gas_unit_price: "100",
+          expiration_timestamp_secs: "1743177404",
+          payload: {
+            function: "0x1::aptos_account::transfer_coins",
+            type_arguments: ["0xd111::staked_coin::StakedAptos"],
+            arguments: ["0x4e5e", "1500000"],
+            type: "entry_function_payload",
+          },
+          events: [
+            {
+              guid: {
+                creation_number: "11",
+                account_address: address,
+              },
+              sequence_number: "12",
+              type: "0x1::coin::WithdrawEvent",
+              data: {
+                amount: "1500000",
+              },
+            },
+            {
+              guid: {
+                creation_number: "4",
+                account_address: "0x4e5e",
+              },
+              sequence_number: "8",
+              type: "0x1::coin::DepositEvent",
+              data: {
+                amount: "1500000",
+              },
+            },
+            {
+              guid: {
+                creation_number: "0",
+                account_address: "0x0",
+              },
+              sequence_number: "0",
+              type: "0x1::transaction_fee::FeeStatement",
+              data: {
+                execution_gas_units: "6",
+                io_gas_units: "6",
+                storage_fee_octas: "0",
+                storage_fee_refund_octas: "0",
+                total_charge_gas_units: "12",
+              },
+            },
+          ],
+          timestamp: "1743177360481259",
+          type: TransactionResponseType.User,
+          block: {
+            height: 311948147,
+            hash: "0x6d02",
+          },
+        },
+        {
+          version: "2532549325",
+          hash: "0x9a6b",
+          state_change_hash: "0xa424",
+          event_root_hash: "0x0321",
+          state_checkpoint_hash: null,
+          gas_used: "12",
+          success: true,
+          vm_status: "Executed successfully",
+          accumulator_root_hash: "0xede9",
+          changes: [
+            {
+              address: "0x4e5e",
+              state_key_hash: "0x3c0c",
+              data: {
+                type: "0x1::coin::CoinStore<0xd111::staked_coin::StakedAptos>",
+                data: {
+                  coin: {
+                    value: "2500000",
+                  },
+                  deposit_events: {
+                    counter: "8",
+                    guid: {
+                      id: {
+                        addr: "0x4e5e",
+                        creation_num: "4",
+                      },
+                    },
+                  },
+                  frozen: false,
+                  withdraw_events: {
+                    counter: "6",
+                    guid: {
+                      id: {
+                        addr: "0x4e5e",
+                        creation_num: "5",
+                      },
+                    },
+                  },
+                },
+              },
+              type: "write_resource",
+            },
+            {
+              address: address,
+              state_key_hash: "0x1709",
+              data: {
+                type: "0x1::coin::CoinStore<0x1::aptos_coin::AptosCoin>",
+                data: {
+                  coin: {
+                    value: "68255318",
+                  },
+                  deposit_events: {
+                    counter: "46",
+                    guid: {
+                      id: {
+                        addr: address,
+                        creation_num: "2",
+                      },
+                    },
+                  },
+                  frozen: false,
+                  withdraw_events: {
+                    counter: "89",
+                    guid: {
+                      id: {
+                        addr: address,
+                        creation_num: "3",
+                      },
+                    },
+                  },
+                },
+              },
+              type: "write_resource",
+            },
+            {
+              address: address,
+              state_key_hash: "0x5520",
+              data: {
+                type: "0x1::coin::CoinStore<0xd111::staked_coin::StakedAptos>",
+                data: {
+                  coin: {
+                    value: "2500000",
+                  },
+                  deposit_events: {
+                    counter: "7",
+                    guid: {
+                      id: {
+                        addr: address,
+                        creation_num: "10",
+                      },
+                    },
+                  },
+                  frozen: false,
+                  withdraw_events: {
+                    counter: "12",
+                    guid: {
+                      id: {
+                        addr: address,
+                        creation_num: "11",
+                      },
+                    },
+                  },
+                },
+              },
+              type: "write_resource",
+            },
+            {
+              address: address,
+              state_key_hash: "0x6f1e",
+              data: {
+                type: "0x1::account::Account",
+                data: {
+                  authentication_key: address,
+                  coin_register_events: {
+                    counter: "5",
+                    guid: {
+                      id: {
+                        addr: address,
+                        creation_num: "0",
+                      },
+                    },
+                  },
+                  guid_creation_num: "12",
+                  key_rotation_events: {
+                    counter: "0",
+                    guid: {
+                      id: {
+                        addr: address,
+                        creation_num: "1",
+                      },
+                    },
+                  },
+                  rotation_capability_offer: {
+                    for: {
+                      vec: [],
+                    },
+                  },
+                  sequence_number: "121",
+                  signer_capability_offer: {
+                    for: {
+                      vec: [],
+                    },
+                  },
+                },
+              },
+              type: "write_resource",
+            },
+            {
+              state_key_hash: "0x6e4b",
+              handle: "0x1b85",
+              key: "0x0619",
+              value: "0xe86e0039581497010000000000000000",
+              type: "write_table_item",
+            },
+          ],
+          sender: address,
+          sequence_number: "120",
+          max_gas_amount: "12",
+          gas_unit_price: "100",
+          expiration_timestamp_secs: "1743176706",
+          payload: {
+            function: "0x1::aptos_account::transfer_coins",
+            type_arguments: ["0xd111::staked_coin::StakedAptos"],
+            arguments: ["0x4e5e", "2500000"],
+            type: "entry_function_payload",
+          },
+          events: [
+            {
+              guid: {
+                creation_number: "11",
+                account_address: address,
+              },
+              sequence_number: "11",
+              type: "0x1::coin::WithdrawEvent",
+              data: {
+                amount: "2500000",
+              },
+            },
+            {
+              guid: {
+                creation_number: "4",
+                account_address: "0x4e5e",
+              },
+              sequence_number: "7",
+              type: "0x1::coin::DepositEvent",
+              data: {
+                amount: "2500000",
+              },
+            },
+            {
+              guid: {
+                creation_number: "0",
+                account_address: "0x0",
+              },
+              sequence_number: "0",
+              type: "0x1::transaction_fee::FeeStatement",
+              data: {
+                execution_gas_units: "6",
+                io_gas_units: "6",
+                storage_fee_octas: "0",
+                storage_fee_refund_octas: "0",
+                total_charge_gas_units: "12",
+              },
+            },
+          ],
+          timestamp: "1743176594693251",
+          type: TransactionResponseType.User,
+          block: {
+            height: 311942427,
+            hash: "0x8655",
+          },
+        },
+      ];
+
+      const transactions: AptosTransaction[] = txs;
+
+      api.getAccountInfo = jest.fn().mockResolvedValue({ transactions });
+
+      const ops = await api.listOperations(address, pagination);
+
+      expect(ops[0]).toHaveLength(2);
+      expect(ops[1]).toBe("");
     });
   });
 });
