+++ libs/coin-modules/coin-aptos/src/logic/isTestnet.ts
@@ -1,0 +1,5 @@
+import { getCryptoCurrencyById } from "@ledgerhq/cryptoassets/currencies";
+
+export function isTestnet(currencyId: string): boolean {
+  return !!getCryptoCurrencyById(currencyId).isTestnetFor;
+}
