+++ libs/coin-modules/coin-aptos/src/logic/isWriteSetChangeWriteResource.ts
@@ -1,0 +1,7 @@
+import { WriteSetChange, WriteSetChangeWriteResource } from "@aptos-labs/ts-sdk";
+
+export function isWriteSetChangeWriteResource(
+  change: WriteSetChange,
+): change is WriteSetChangeWriteResource {
+  return (change as WriteSetChangeWriteResource).data !== undefined;
+}
